+++ user-api/src/main/java/com/biengual/userapi/learning/presentation/dto/UpdateLearningRateDto.java
@@ -1,9 +1,11 @@
 package com.biengual.userapi.learning.presentation.dto;
 
+import java.math.BigDecimal;
+
 public class UpdateLearningRateDto {
     public record Request(
        Long contentId,
-       Integer learningRate
+       BigDecimal learningRate
     ) {
     }
 }
